Add new endpoints for RotateKey and UpgradeKvm
Moved the functionality for upgrading KVM and key rotation from
management calls to new proxy endpoints. EMG 3.1.5+ will need to be
upgraded to a minimum of this version of the edgeauth proxy to function
correctly.
The new endpoints are /rotateKey and /upgradeKvm, and will be called by the
corresponding edgemicro cli commands